Day 24, part 2 done!
This one was really cool and really tricky.  I wrote code
(circuit-fix.py) to rename the wires to represent what part
of the sum they were computing.  Then I found the wires that
didn't get renamed because they didn't fit the patterns

The part2.py program evaluates the adder and find which bit
position it breaks on first.  With those two I was able to find
the mistakes and fix them until the adder worked.

There's maybe a better way, but this was effective enough.